Add an aliasing feature
In the Solidus ecosystem, the user class is configurable, and many
extensions patch the user class using a pattern like
`Spree.user_class.prepend`. `Spree.user_class` is not a string, and
can't be constantized, so we need to replace it with the actual class
name before patching. This commit accomplishes that.